Added tx (and later more) retrievers
Using public API's (Blockr's to be specific) over server proxy.

Ideally we'd have the client do this, OR maintain the blockchain
ourselves on the server (which is meh, because that's costly, but the
only way to serve reliable data we ourselves feel confident about)

Not even sure if there are any API's out there that support jsonp,
although I suspect they all do, as they should.

TODO: Figure out a way to be able to serve reliable data out of the
blockchain without using any resources on the server (in fact, I'd like
the server to not have to run in a servlet container, or perform any
logic at all, client only)