Refactored a whole lot of UI
- Added coinbase tx to block view
- Refactored HexViewer to be replaceable
- Server calls now add the coinbase tx info and the raw tx to the
returned value

---

Need to look at the performance of HexViewer some time, it's beginning
to look like a blown up turd with all the recent changes.

The genesis block coinbase transaction is not returned by RPC (because
it is an unspendable, / invalid tx). Which is crap. Disgusting
workaround on the server end which may need some attention.

Getting block information is currently doing 3 rpc calls, 4 if you
include the initial raw block call. These calls are doubly performed (2x
getblock and 2x getrawtransaction). The latter is easy to fix, the
former not so much. The architecture for this call is the same for the
tx page, which is also doing the same call twice. Maybe we'll have to
look into bundling these calls (for example, do we really need to get
the raw transaction string seperate from the blockchain object? -- same
goes for blocks)