Fixed dependency
This project is using a slightly modified version of gwt-material, a version which excludes the inheritance to the Clean theme, and without various other styling overrides. These modifications previously had to be built manually, as they do not exist on a nexus anywhere. The pre-built project has now been added to a github repo, and added as a dependency repository. It's no longer required to manually build the modified gwt-material project as a result, although the build process has been made a bit messier - hopefully gwt-material will exclude the default inheritance of the Clean theme so this is no longer needed - request to doing this is outstanding.